requirements.txt
mcp[cli]>=1.0.0
streamlit>=1.32.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pandas>=2.0.0
plotly>=5.0.0